The first task executor fails on scales different from `1`
I used `h * WSExecutorFirst.BASIC_HEIGHT + i`
for the final digit matrix row index,
where `i` is the ideal digit row index
and `h` is a number of the current row repetition.
The code contains a mistake,
because I've written it like I want to index a matrix
by a 1D array.
The correct code is `i * this.verticalScale + h`:
there are `vertical scale` rows of the final digit
from one row to its neighbor of the ideal digit.